+++ src/main/java/subway/ui/handler/ErrorResponse.java
@@ -1,0 +1,14 @@
+package subway.ui.handler;
+
+public class ErrorResponse {
+
+    private final String message;
+
+    public ErrorResponse(final String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
